Prevent repo size from ballooning with static data
And in the process, ensure scripts allow re-creation of the
data that is now in gitignore. Ideally we'll move all this data
to some storage bucket and wire up a deployment process on github
but for now, french data stays in and serves as the test bed on
firebase preview channels, the others will just stay local, and
deployments will be done via the firebase hosting CLI manually.